Tpetra: computeOffsetsFromCounts can now take a host View for counts
@trilinos/tpetra Tpetra::Details::computeOffsetsFromCounts can now
take the input View of counts in a different memory space (e.g., host
vs. device) than the output View of offsets.  This serves a common use
case in Tpetra::CrsGraph and Tpetra::CrsMatrix, where counts live on
host but offsets need to go on device.

The unit test now explicitly exercises this use case, as well as the
case where both counts and offsets live in the same memory space.  The
test also checks the return value (the total count) of
computeOffsetsFromCounts.