stk-eval: fix segfault if function is not found but still evaluated
Fix for ticket #15209

User output in an Adagio simulation included the expression
"sqrt(Dinp(1)^2 + Dinp(2)^2 + Dinp(3)^2)". The expression evaluator
interprets Dinp as a function which does not exist (expected
behavior). An undefinedFunction flag is set to true, but the list of
undefined functions is checked against empty in order to throw an
error. The function is now added to the list which throws an exception
and emits a helpful message.

Updated unit test to account for thrown exception.

(From STK/SEACAS)