fix delegation in Piro::TrapezoidDecorator
This decorator was deriving straight from
Thyra::ModelEvaluatorDefaultBase and doing
a lot of its own delegation, some of which
was not quite right.
Specifically, when delegating createInArgs
and createOutArgsImpl, the argument object's
model description needs to be changed,
at least to comply with checks enabled when
Teuchos_ENABLE_DEBUG=ON.
Thyra has a base class ModelEvaluatorDelegatorBase
specifically for creating decorators, so use
that as the base class and let it provide
most of the delegating methods.

This class seems to only be used by Albany,
and this commit allows an Albany test to
satisfy certain Thyra safety checks that it
did not before.