Zoltan2: Converts tests to run as an internal CMake loop. #496
Originally we were going to have different Jenkins tests.
We decided it would be better to have it all in a single test.
This runs a very extensive loop.
It also has a setup for defining individual tests.
From discussion it seems we may just define them that way.
So this this is still in progress.