IOSS: Remove ifdefs around chdir kluge for long filenames.
Many mpi implementations seem to be based on the mpich
mpiio code which has the buffer overflow.  It has proven
difficult to have a concise ifdef for applying the kluge
only to the buggy mpi versions.  Therefore, we always do
the chdir kluge.  Maybe in several years we can remove
the kluge; however, it should not be very inefficient
to always do this call.

If this shows up as a hot-spot on some filesystem, then
check the filename length and only do the chdir if the
length is over 240 characters...

(From STK/SEACAS)