MueLu: bugfix in RebalanceBlockAcFactory
The factory now can deal with empty blocks (i.e., Teuchos::null) in the blocked operator.
However, we still implicitly assume that the diagonal blocks are all non-null, i.e., they
can be empty but they are initialized as proper CSR matrix blocks (which may or may not be
the case in general). We're going to fix that later.