Kokkos: Fix warning in DynRankView
This commit fixes a warning in the DynRankView. This will fix the
failing Trilinos "warnings as error" builds for phalanx. This same fix
is already accepted into the kokkos develop branch. There is no
timetable on the next sync of kokkos into trilinos so this temporarily
fixes the warnings. The next snapshot of kokkos will overwrite this
commit with the same fix.

kokkos issue #381