Tpetra: added check on #rows in exception
An extra condition has been added in the exception checking that the
LU factors be actually lower triangular and upper triangular inside
the trsv method.  This was needed because the exception is thrown when
one processor does not own any row in the RowMap associated with a
Tpetra::CrsMatrix object. Now the exception also checks that #rows
on the node is not zero.