Disable optional connectivity-map capability in preparation for removing it.
Its utility in saving memory was never realized, and it is not used by
any known application (verified to not be used by Nalu, Aero, Albany or
framework-based apps). It is a sustainability problem in BulkData, and
even a performance problem in some scenarios due to the if-else
statements that get connectivity two different ways depending on whether
a minimal connectivity-map has been specified.

(From STK/SEACAS)