Zoltan2: Initial cdash testing setup for zoltan2.
Adds the TrilinosCTestDriverCore for jenkins testing.
Used a CUSTOM_ENV_VARIABLES flag to define places where that
file has custom changes for the initial testing and setup.

Adds the Zoltan2 script for running a single Jenkins job.
Currently we are testing these by setting the Jenkins parameters
in a script and the calling the ctest script.