Intrepid: Fix warning in MiniTensor
Changes provided by Alejandro Mota

Added workaround for spurious warning in gcc 4.x.x.
Added workaround for failure to recognize function template specialization in gcc 4.7.2.

gcc 4.7.2 emits a spurious “array subscript is above array bounds” when compiling with
-Werror=array-bounds. See

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=45180
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53198
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=59124

As a workaround introduce a guard for gcc 4.x.x in the indexing operation triggering
this warning and use a conditional statement to force that compiler to realize that
the index is within bounds.

gcc 4.7.2 fails to recognize full function template specialization and thus produces
duplicate symbol link errors as it takes the specializations as separate functions.
These specializations were introduced in the number_digits utility function template
for use in CUDA as CUDA does not accept many functions and classes in the std namespace.
Since Intrepid is not being converted for use with CUDA (as opposed to Intrepid2),
revert to the use of std for the number_digits function template.

Build/Test Cases Summary
Enabled Packages: Intrepid
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=150,notpassed=0 (4.96 min)
1) SERIAL_RELEASE => passed: passed=150,notpassed=0 (1.89 min)