ml: new ML option               smoother: min size for coarse xxx
This allows users to indicate a size that the coarse level must be in order for the coarse smoother to be applied. If above this size, the fine level smoother is instead invoked. This address an issue that arose in an ARIA run where because of threshold dropping the entire fine grid matrix was viewed as a diagonal operator. This caused the coarsening to abort, which then cause a direct solver to be applied to the entire fine level problem (on one processor). Now, the fine smoother can instead be applied once coarsening aborts.