Intrepid MiniTensor: Fix warning in MiniTensor
gcc 4.7.2 emits a spurious “array subscript is above array bounds”
when compiling with -Werror=array-bounds. See

    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=45180
    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53198
    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=59124

As a workaround introduce a guard for gcc 4.x.x in the indexing operation
triggering this warning to force that compiler to realize that the index
is within bounds.

This is needed because Sierra takes compiler warnings as errors.

Build/Test Cases Summary
Enabled Packages: Intrepid
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=150,notpassed=0 (10.60 min)
1) SERIAL_RELEASE => passed: passed=150,notpassed=0 (4.00 min)