Tpetra::{Row,Crs}Matrix: Added new getLocalRowViewRaw method
@trilinos/tpetra Tpetra::RowMatrix has a new virtual method,
getLocalRowViewRaw.  It is the "raw pointer" version of
getLocalRowView (that uses Teuchos::ArrayView).

The default implementation of getLocalRowViewRaw just calls
getLocalRowView.  Tpetra::CrsMatrix's implementation of
getLocalRowViewRaw bypasses Teuchos::Array* data structures
completely.  This avoids the known thread safety issues of
Teuchos::Array* data structures, especially in a debug build.  As a
result, it is safe to call CrsMatrix's implementation of
getLocalRowViewRaw in a (non-CUDA) thread-parallel context.