Ifpack2: Fix #558.
The new LocalSparseTriangularSolver is handling a more general case than the
original bare localSolve's that were used previously in RILUK and ILUT. But this
generality has uncovered a bug in LocalFilter and OverlappingRowMatrix
(depending on how one thinks of it) when AdditiveSchwarz is used with overlap >
0.

LocalFilter is intended to construct local domain, range, column, and row maps
based on the input matrix's domain and range maps. In fact, it was using the
input matrix's row map for the range map and the input matrix's domain map for
the domain map, mixing these. This commit is based ont the premise that this was
a fix for another bug long ago that is similar to the current bug, except in the
range/row maps rather than (now) in the domain/col maps.

This commit takes a different approach. It returns LocalFilter to using *just*
the input matrix's domain and range maps. Then it makes OverlappingRowMatrix
return RowMap_ as both row and range maps, and ColMap_ as both column and domain
maps. The idea here is that OverlappingRowMatrix's intent by construction must
make the input matrix's domain and range maps no longer relevant. Then when
LocalFilter takes OverlappingRowMatrix, it gets domain and range maps that make
sense.

The broken case reported in #558 was verified to be fixed.

Add a test to distinguish between broken (previously) and fixed (now).

Also remove redundant L_ and U_ fillCompletes from RILUK.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=68,notpassed=0 (39.36 min)
1) SERIAL_RELEASE => passed: passed=63,notpassed=0 (30.25 min)