Intrepid2 MiniTensor: Change output from tests. If no arguments are given in the command line, minimal output is provided (essentially just the test name and PASSED or FAILED). If more than zero arguments are provided, then print information useful for tracing what the tests are doing and for debugging.
Also change class for equality constraints so that both the number
of variables and numbers of constraints are provided. Previously
the number of variables was deduced from the capacity of the
vector of variables passed to the value/gradient functions. The
change is so that vectors of arbitrary capacity can be supplied to
the value/gradient functions, as long as their dimension is equal
to the number of variables specified in the constraint class.

ROL: Propagate the above changes.

Build/Test Cases Summary
Enabled Packages: Intrepid2, ROL
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) SERIAL_RELEASE_ST => passed: passed=242,notpassed=0 (20.17 min)