KokkosKernels: Only pre-build multi-RHS mat-vec for default offset type
KokkosKernels was pre-building the multiple right-hand sides version
of sparse matrix-vector multiply for the explicit offset type size_t.
Since Kokkos::Cuda may not necessarily use that as its default offset
type, this may have pre-built a lot of code that never gets used.
This commit makes sure that KokkosKernels only pre-builds this kernel
for the default offset type.

This relates to #700.  Build size of Trilinos/packages/tpetra for
MPI_RELEASE (dynamic libraries) doesn't change much in a non-CUDA
build as a result of this commit, but it's good to be sure.