Update Scotch TPL test
The test for version number on Scotch was failing to compile with version
6.0.3. This appears to be an issue with Scotch not including a header,
stdint.h, necessary for its own use. Since this version of Scotch is the
version that is being checked for and since this version of Scotch works
with all the packages that use it I adjusted the test for the version to
include the missing header so that it will compile.