Remove a dependency for finding UMFPACK
UMFPACK was requiring that UFconfig.h be present. This file does not
exist in newer versions of SuiteSparse which contains UMFPACK and was
occasionally, but not always, causing the configure to fail looking for
the file. Since the correct configure file changes with the verison
and is part of the install it is reasonably safe to assume that if
there is an umfpack.h and amd.h in the suitesparse then the correct
config.h is also present.