Adjust for OSX: GCC 5.3.0, Boost 1.58.0, no Scotch (#158, #482)
The SEMS Env does not provide any GCC except for 5.3.0.  Also, it does not
provide Boolst 1.55.0, only 1.58.0 and 1.59.0.  Also, it does not provide any
build of Scotch at all.

Therefore, I have change the default dev env to be GCC 5.3.0, Boost 1.58.0 and
disabled the load of Scotch.

Currently need to manually set TRILINOS_DIR.  Need to fix this up so that it
is more automatic.  See comment in the script.