Panzer: Split up Owned and Ghosted Vectors (#615)
This is an initial attempt at splitting the x_ vector in
GatherSolution_Epetra into xOwned_ and xGhosted_.  If we're dealing with
a new EpetraVector_ReadOnly_GlobalEvaluationData, we'll use the new
owned and ghosted vectors.  If we're dealing with an older
EpetraLinearObjContainer, we'll deal with x_ as before.

All UniqueGlobalIndexers were modified such that they now have the
following routines:
* getOwnedIndices()
* getGhostedIndices()
* getOwnedAndGhostedIndices()
* getNumOwned()
* getNumGhosted()
* getNumOwnedAndGhosted()

The Epetra_LinearObjFactory was modified to build the ghosted map using
only the ghosted indices (as opposed to including the owned ones).
EpetraVector_ReadOnly_GlobalEvaluationData shouldn't need any changes,
as it'll build the vectors based on the owned and ghosted maps.

THIS CHANGE IS NOT READY TO PUSH!  There are about a third of our
(Panzer, Drekar, and Charon) tests that fail or timeout for some reason.
Still trying to track that down.