Remove all non-SEMS env enables/disables (#158, #482)
This commit removes all of the enables/disables from this file.  Now the
SEMSDEvEnv.cmake module just sets the locations for compilers and the TPLs.

What should get enabled and disabled for CI testing is really orthogonal and
should be set somewhere else.

Also, I have realized due to many reasons that we may not be able to able to
demand the SEMS env in order to define a CI build.  If someone has a
compatible set of compilers, MPI, and TPLs built, then perhaps we should allow
them to push?  That will be will supported when the checkin-test.py script
supports --compare-to-control-build option (see TriBITSPub/TriBITS#152).