Replace deprecated PS,SS with PT,ST (#482)
The names "Primary Stable" PS and "Secondary Stable" SS are deprecated and
instead have been replaced with "Primary Tested" PT and "Secondary Tested" ST
to differentiate them from maturity levels (see the TriBITS Lifecycle Model).

I also put in the function TRIBITS_REPOSITORY_DEFINE_TPLS() in TPLsList.cmake
and took out the documentation.  Now people can find the documentation for
this function in the TriBITS Developers Guide.

I did this refactoring to get ready for the change in the set of PT and ST
packages and TPLs as part of trilinos/Trilinos#482.