Attempt to fix python issues on continuous build
It appears that tribits has been prefering python 2.6 since it
changed over to using cmake's findPythonInterp module. We have been
seeing several python related issues with no explanation why. This
attempt will cause tribits to prefer 2.7 which is generally what we
have installed. If this works we will need to find a better more
permanent solution for this.