Panzer:  Split up Owned and Ghosted Vectors (#615)
Eventually figured out the problem was something to do with
Thyra::get_Epetra_Vector().  It's like there was something from that
call getting attached to the RCP that was then being problematic when it
came to the compiler-generated destructor for GatherSolution_Epetra.
The fix at this point was to grab the ownedMap_ from the
EpetraVector_ReadOnly_GlobalEvaluadtionData and store it as member data
in GatherSolution_Epetra.  Then we can keep xOwned_ as a
Thyra::VectorBase<double>, and when we need to access its memebers via
local IDs, we can create the appropriate Epetra_Vector from it.

Anyhow, currently xGhosted_ still contains everything, and x_ is still
being set alongside xOwned_ and xGhosted_.