MueLu: EpetraOperator returns the merged full maps for Operator{Range,Domain}Map
The Epetra interface is directly used with AztecOO and has no native support
for product MultiVectors as they are included e.g. in Thyra. Therefore, we
have to use the fully merged single-block multi-vectors that are splitted again
in Xpetra. For efficiency reasons you might consider to use the Thyra interface
with Belos as linear solver instead.