Domi: Add more sophisticated processor decomposition algorithm
This commit adds a much smarter algorithm for decomposing the
processors along each axis of an MDComm. Previously, if there was
more than one axis with unspecified comm dimensions, the first axis
would get all of the remaining processors, and the rest would get one
processor.

Now, it is possible to to get a much more logical decomposition, and
one that takes the dimensions of the problem into consideration. For
example, consider a field with dimensions {10,40,20}. The best way to
decompose this for 8 processors is {1,4,2} which would result in 8
10x10x10 blocks. The old way would have broken it down {8,1,1} where
the 8 processors would have been dividing up 10 grid points along the
x-axis.

Users only get this capability when using the ParameterList
constructors within Domi, where the "dimensions" parameter can be
specified or not as the user sees fit. (If this parameter is not
specified, then each axis gets the same weight.)

To add this new functionality, new functions were added to the
Domi_Utils files, the MDComm constructors that use ParameterLists were
modified to use the new functions, and the MDMap and MDVector unit
tests that assumed the old decomposition had to be re-written to
account for the new decompositions.