TestScripts: add test_script to compare two repositories
This script will take to source locations for Trilinos and a
config_script as input which uses TRILINOS_PATH as the source location.

The script will the configure using the first repositories, build and run.
If any errors occur during build or run, it will proceed
to configure build and run based of the second repositories.

It will then compare the errors of the two test cycles. If they agree
the script passes, if they don't it will print out all the differences
and exit as failed.

This can be used to test changes on Trilinos if Trilinos is not in a clean state.
The script is designed to use lsf (bsub) to submit configure, compiler and test
jobs.