Add MiniTensor TriBITS Packages to Trilinos TriBITS Repo (#609)
A TriBITS package does not use a ProjectName.cmake, PackagesList.cmake, or
TPLsList.cmake file.  ProjectName.cmake is for a TriBITS Project (i.e. a CMake
project you directly configure).  PackagesList.cmake and TPLsList.cmake are
for a TriBITS Repository (which MiniTensor is not).

With this, MiniTenor configures correctly under Trilinos.