Tpetra: Let iallreduce take in & out Views of different layouts
@trilinos/tpetra This will help with #957.
Tpetra::Details::iallreduce used to require that both the input and
the output buffers have the same array_layout.  This was OK for rank-1
Views, because Kokkos allows assignment of LayoutLeft to LayoutRight
or vice versa for rank-1 Views.  However, Kokkos does NOT (currently)
allow assignment of LayoutLeft to LayoutRight or vice versa for rank-0
Views, even though this would make sense.  See the outstanding Kokkos
issue:

https://github.com/kokkos/kokkos/issues/594

This commit adds template parameters for the array layouts of the send
and receive buffers, to iallreduce.  This lets the send and receive
buffers have different layouts.  In theory, only LayoutLeft and
LayoutRight should work, though we do not currently check this.  (I'm
not sure if LayoutStride makes sense for a rank-0 View.)