PyTrilinos: Update FindNumPy to work with debug version of Python
The debug version of Python can output the total number of references
to standard error from time to time, and this was messing up the logic
in FindNumPy.cmake. This commit fixes that particular use case without
affecting the normal use case.

Not that this logic is correct, although I didn't get building
PyTrilinos with a debug version of Python to ultimately work.