PyTrilinos: Fix tests with MPI_Finalize() problems
It appears that if a module external to Epetra (such as TriUtils or
Galeri) constructs Epetra objects, those objects can get destructed
after the Epetra module. This is a problem because the Epetra module
can call MPI_Finalize(), and the Epetra objects' constructors assume
this is not the case:

*** The MPI_comm_size() function was called after MPI_FINALIZE was invoked.
*** This is disallowed by the MPI standard.
*** Your MPI job will now abort.

This only happens after the script has completed, so it is not
strictly a problem, but it looks bad. As I workaround, I modified
these test scripts to "del" the offending object near the end of the
script. It would be better if Epetra_MPI_Finalize() could be smart
enough to delete these objects on its own.