zoltan2:  fixed use of Tpetra Map's IndexBase.
Some Zoltan2 code was using IndexBase (sort of, but not quite)
like the minimum LID on a processor, when in truth, IndexBase is
just an index to the start of raw data in arrays input to the
matrix or map constructor.  It is meant to differentiate between
Fortran and C/C++ arrays.

The Zoltan2 library always accesses matrix data from the Tpetra
accessors, not through the raw pointers.  Thus, it doesn't need
to fuss with IndexBase.  And the way it was using it was wrong.

Thanks to @vbrunini for identifying this issue. #973

This commit includes the start of a test of non-contiguous matrix IDs,
but the test is not yet enabled.