Zoltan2: Move casting functions out of factories - resolve a few errors.
Moves the utility functions for safe upcasting out of the factories
and into test_driver or ComparisonHelper. This makes the factories
(ProblemFactory, EvaluateFactory, and AdapterFactory) not contain this
complication. A few methods which don't depend on the Adapter type can
be placed in the root class but this will need further discussion.
I made this change in BaseAdapterRoot to demonstrate for
getLocalNumIDs and getNumWeightsPerID().

Fixed an error in local/global code for currently disabled test AlgND code.
Fixed a duplicate header.