Tempus: Fix UMR for DIRK Stepper.
The initial stage x was uninitialized, which is also the initial
guess to the nonlinear solver for the first stage of the first
timestep.  Most of the time NOX was able to still get a solution,
but occasionally would fail on large timesteps.

Also added the van der Pol problem for the DIRK stepper.