Amesos2: Implement Umfpack wrappers for Amesos2
The wrappers are modeled on the SuperLu wrappers.
The tests also closely follow the SuperLu format.

One important difference is Umfpack supports only
double types, not float types. Conversions have been
implemented and a warning is given in CMake if
float is on and double is off. However some tests
don't currently build for Tpetra under those conditions.

Note that linking to an external suite sparse
can create warnings due to namespace collisions.
A separate PR will resolve this.