Zoltan2: Environment change resetParameters and all constructors have comm #1014 #1004
Changed the way resetParameters() is setup as discussed.
This removed the old change and implemented a new one.
This is not expected to change behavior.

Turns on rcb_C unit test always.

Changed default Environment constructor to take a comm.
Changed all code dependent on default Environment().
Generally this meant:
  CHanging a lot of main() functions to simply pass comm.
  Changing a few error messages to be handled manually instead of with Env.

Note that in Zoltan2_TaskMapping.h I changed an API to take an
Environment as I believe the problemComm should not be used in this place.
This is called from taskMappingTest3.cpp.
However I am not sure if this is desirable or even ok to change that API.