Newmark Beta: refactoring code so that there is a separate Model Evaluator for second order schemes like Newmark Beta from the original Tempus::ResidualModelEvaluator (for 1st order schemes).  This is a nicer design, as 2nd order time integrators will be fundamentally different from first ordere ones.
Moving xdotdot-related logic to newly created Tempus::SecondOrderModelEvaluator class.

Creating test for Newmark Beta scheme involving throwing a ball up in the air
with a parabolic solution, to be used to debug Newmark Beta scheme
as it is implemented.  This code can be viewed as a shell for now -- needs to be filled in (WIP).