Small clean up to orthogonalization parameter handling.
I noticed, while looking into bug #965, that the current orthogonalization
parameter being used by any of the Belos solver managers was not stored in the
"current parameters" list.  This corrects that issue.

Also, I noticed that OrthoManagers were private members of several
solvers that didn't use them.  So, I cleaned that up by removing the
OrthoManager that was not being used.