Fix up Stratimikos config files (#1023)
The file src/Stratimikos_Config.h.in was never used.  It must have been left
over from the autotools days.  Very sloppy (on my part likely).

I had to put these defines in the Stratimikos_Config.h.in file because that is
needed for clients to pull in the right headers and maintain backward
compatibility.