Updated Epetra examples for BlockDavidson and LOBPCG, added new SI example for BKS
Not all of the examples for BlockDavidson and LOBPCG converge with the solver
settings in the example.  That's not very instructive.  I've gone through the
examples, updated the parameters (number of eigenvalues, blocksize, sort, etc.),
and they converge on my laptop.  That does not apply to the examples that read
in Matrix Market matrices, because convergence is dependent upon the matrices
being read in.  I've also added a new BKS example that will read in a Matrix
Market matrices and create an eigenproblem that is solved using a transformed
operator (inv(K)*M).