Fixing error is logic with solve() is called without factor() with equilibration.
As pointed out in Github issue #1034, there is a flaw in the logic of
the solve() method in the DenseSolver classes.  There is a check that both
the right-hand side vector and the matrix has been equilibrated, when
equilibration is enabled, before the check if the matrix has been factored.
If the matrix has not been factored, then equilibration has not been applied
yet, thus this check fails and the method() returns an error.  This flawed
logic has been cut-and-paste into all the DenseSolver classes.  So, all of
these classes needed to be fixed.

The DenseSolver tests have been updated to check that this logic has been
corrected, by testing this exact scenario.  Furthermore, in writing these
new tests, an error in the logic of the solveWithTransposeFlag() method was
found.  This error would not allow the user to change the transpose state
of a solver object from "transposed" back to "non-transposed".  The other
solveWithTranspose() method works just fine.  This error, again, was
cut-and-paste into the other DenseSolver classes, which had to be fixed.