EpetraExt: Add HDF5_IO.cpp as a TRIBITS example. (#1062)
This commit follows on PR #1062.

Changes were made to the API for HDF5 but no
automatic testing was in place.

The HDF5_IO.cpp example was added to build as an EpetraExt example.
It was decided that the other examples should be added later
and an issue will be opened for those.