Kokkos-Kernels: Adding Kokkos-Kernels as a stand-alone package
Kokkos-Kernels was moved from its previous location tpetra/kernels
to kokkos-kernels and now is its own package. Testing was performed
on multiple platforms with gcc, intel and cuda, also debug mode.
As far as I can tell the have the same errors as on current develop.
I.e. hopefully nothing related to this switch. Testing was done with
the samplescripts/Sandia-Sems/configure-testbeds-jenkins-all script.