Added all necessary simple-DG code to Panzer for face decomposition in local meshes. Modified STKConnManager to use a const STK_Interface. Added a comm request function to STK_Interface so that we can access a comm designed for the domain the mesh exists on. Function throws error if mesh is not initialized. Added utilities for handling mesh partitioning (nothing is working or tested yet). Added IntegrationDescriptor and BasisDescriptor for identifying IntegrationRule and PureBasis in workset. This interface now exists in WorksetNeeds as well. Added requested workest size to WorksetDescription with default state of EMPTY (enum) for backward compatibility. Other options are FULL or an int >=0. Added requires partitioning flag to WorksetDescription. Flag defaults to false, and should be backward compatible. Changing it to true will branch the workset factory to request things from the partitioned workset utilities, which don't yet work. Added FaceToElement class to allow for finding faces for a given chunk of the mesh. Added local mesh utilities which will define ghost cells and faces for a local chunk of the mesh. Added local mesh chunk interface which will be derived by partitioning the LocalMeshInfo structure.
There are currently deprecation flags throughout the workset interface which will be uncommented when this work is complete.

Things to do:
Test code when GlobalOrdinal is set to 64 bit unsigned. Code supports this, but it's completely untested.
Get the partitioner working. Currently it assumes there is only one parition and throws an error if more partitions are requested.
Get the sideset partitioner working. Basis code outlines how it will be setup (i.e bypasses partitioning step), but needs to be finished.
Write the partition/chunk to workset interface. Should be very simple, but we need to know more about how sidesets will be setup.
Enable the deprecation warnings and fix all the IR and Basis rule setting into Workset so that we can do IR/Basis lookups with the descriptors.