Teuchos: fix typing of TwoDArray::operator[]
This operator must be rarely used, because
the return type doesn't match the type of
the expression returned (it should be
ArrayView<const T>, not const ArrayView<T>).
I'm fixing this because I need this operator
to unit test some YAML functionality I'm
developing.
(I can't use the operator(i, j) because there can
be no commas in the argument to TEST_NOTHROW).