Stokhos/Piro:  Updates to allow Adaptive and Interlaced SG ModelEvaluators
This allows the use of the Adaptive and Interlaced form of the SG
ModelEvaluatos through Piro.  In the Piro "Stochastic Galerkin"
parameter list, you can specify "SG ModelEvaluator Method" to "Default",
"Adaptive" or "Interlaced".  Not that not all funcationality works with
all three.  Also, if the basis is adapted, Albany currently ends with an
exception because there is an inconsistency in the g response
corresponding to the solution.