Mod in Thyra_BelosLinearOpWithSolve: separating verbosity of BelosSolver from that of Belos LOWS
The verbosity of Belos LOWS is set by the Belos LOWS factory to be the same
as that of the Belos LOWS factory.

Before this change, the ostream of the Belos solver was set to be a
oblackholestream if the LOWS verbosity was not at least equal to VERB_LOW,
in which case it would make the 'Verbosity' option in the Belos parameter
list completely ineffective. This, in turn, forced the user to give the
LOWS and the LOWS factory some degree of verbosity in order to allow the
Belos solver to print anything at all.

Since the user is not necessarily interested in the output of the LOWS and
LOWS factory, we modified this. Now the ostream passed to the Belos solver
is a valid ostream (not a blackhole) and the user can set the verbosity of
the Belos solver via parameter list independently of the verbosity of the
LOWS and LOWS factory.