Tpetra::CrsMatrix: Optimize {replace,sumInto}GlobalValues (see #800)
@trilinos/tpetra The versions of replaceGlobalValues and
sumIntoGlobalValues that take "raw" input arrays used to convert them
to Teuchos::ArrayView and then call the versions of those methods that
take Teuchos::ArrayView.  The latter in turn would convert _back_ to
raw arrays and then to Kokkos::View.  This commit removes the
intermediate step of conversion to Teuchos::ArrayView.

This is a small step towards #800.  Thread-parallelizing CrsMatrix
pack and unpack must begin with basic thread safety.  Since
Teuchos::ArrayView is not thread-safe in debug mode (see #229), and
since it will never work on a GPU, the first step is to stop using
Teuchos::ArrayView in pack and unpack.  Instead, we must use pointers
/ Kokkos::View objects all the way through.