Tpetra::CrsMatrix: Add new non-public packRowStatic method
@trilinos/tpetra This is related to #800.  See my comments there, in
particular "we first have to change packRow so that it goes directly
to the KokkosSparse::CrsMatrix if that exists, rather than going
through the 'generic' getLocalRowView / getGlobalRowView interfaces
that return Teuchos::ArrayView instances."  This commit is the first
step to accomplish that subgoal.

Tpetra::CrsMatrix::pack now uses packRowStatic when the graph is
static.  Otherwise, it falls back to packRow.  This ensures that the
new method gets tested.