Fetch and merge in branch on remote before running checkin-test-sems.sh (#482)
Turns out that are use cases where you really need to fetch and mege in the
branch *before* running checkin-test-sems.sh on the remote machine.  For
example, there was a use case where we needed an updated PackagesList.cmake
file in order to elevate a package (Tempus) from ST to PT so that the
checkin-test.py script would enable it.  But if the checking-test.py script
runs before that change, then it does not see that upgrade from ST to PT and
therefore will not enable the package, and will not push.

Now the checkin-test-sems.sh script will run on the remote machine seeing all
of the changes, which means it will run identically to if it were run locally
(only it is more robust because it will not get tripped up by untracked
ignored files that could be needed by the build or tests).

These were use cases that I had not considered before but they are important
use cases for better robustness.