Update Ride build to reflect module changes on the system
A number of modules that we were loading were hard coded in the scripts and
the module no longer exists. In addition it looks like specific blas and lapack
modules no longer exits so I removed the usage of them for this in the hopes
that the default logic will figure it out.