Add a blocking mode to address SSH problems (#482)
Now you can pass in the third argument as 'blocking'.  This will result in a
single blocking remote SSH invocation.  The advantage of this appraoch is that
you can just type your password once for the remote machine in case you can't
get your public/private key to work.  The disadvantage is that you will see a
lot of output about what is getting built and tested.  (I really need to
reduce the amount of output that checkin-test.py sends to stdout and leave the
checkin-test.out file as a more detailed log file).