Teuchos: ensure scalar types are preserved
This commit solves two issues, both of which
fall into the category of "the type of a
parameter would change if written to YAML
and read back into Teuchos":
1. numbers with quotes around them would
   not be interpreted as strings
   (which is necessary to ensure strings
    remain string-typed even if they happen
    to just contain a number)
2. the API that takes only a filename
   would print integer-valued doubles
   in integer format (YAML has no other way
   of distinguishing double and int, so I
   added std::scientific to ensure all doubles
   get formatted differently from ints)
While fixing item (2), I added the precision
settings discussed in #1181.
There is no way for users to change those settings
from the API that takes only a file name, so
high precision seems like a good default.