Sacado:  replace usage of ternary operator with if_then() function.
Replaced usage of the ternary operator "cond ? a : b" with a new
function "if_then(cond,a,b)".  Sacado provides a default definition of
if_then() that will work for most scalar types.  This is necessary for
Sacado to wrap the stk::simd::Doubles scalar type for Sacado usage in
Sierra, where conditionals involving that scalar type don't return bool
(but rather a simd bool).  Such a concept could also be useful for my
ensemble scalar type in Stokhos.

For the moment I've only done the overloaded operators in the
Sacado::Fad namespace.  After verification this works as expected in
Sierra, I will propagate the changes to the other Fad namespaces.